bot/requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
google-generativeai==0.8.2
pydantic-settings==2.5.2
